move metrics & recovery middleware to own packages
