+++ internal/server/middleware/recovery/recovery.go
@@ -1,4 +1,4 @@
-package middleware
+package recovery
 
 import (
 	"context"
@@ -8,7 +8,6 @@ import (
 
 	"github.com/go-kratos/kratos/v2/log"
 	"github.com/go-kratos/kratos/v2/middleware/recovery"
-	"go.opentelemetry.io/otel/metric"
 
 	"google.golang.org/grpc"
 )
@@ -16,9 +15,12 @@ import (
 type Option func(*options)
 
 type options struct { // Duplicated from https://github.com/go-kratos/kratos/blob/main/middleware/recovery/recovery.go b/c no export
-	handler  recovery.HandlerFunc
-	requests metric.Int64Counter
-	seconds  metric.Float64Histogram
+	handler recovery.HandlerFunc
+}
+
+type requestInterceptingWrapper struct {
+	req any
+	grpc.ServerStream
 }
 
 func StreamRecoveryInterceptor(logger log.Logger, opts ...Option) grpc.StreamServerInterceptor {
